Playground changes for trait obj Eq
